fix some bugs regarding to milestone 2

+++ src/main/java/com/flipmind/localizationservice/controllers/TranslationController.java
@@ -1,6 +1,7 @@
 package com.flipmind.localizationservice.controllers;
 
 import java.io.IOException;
+import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
 
@@ -16,12 +17,14 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.flipmind.localizationservice.GlobalVariable;
 import com.flipmind.localizationservice.models.Document;
 import com.flipmind.localizationservice.models.Locale;
 import com.flipmind.localizationservice.models.Project;
+import com.flipmind.localizationservice.models.Status;
 import com.flipmind.localizationservice.models.Tenant;
 import com.flipmind.localizationservice.models.TranslatedString;
 import com.flipmind.localizationservice.models.Translation;
@@ -32,6 +35,7 @@ import com.flipmind.localizationservice.repositories.LocaleRepository;
 import com.flipmind.localizationservice.repositories.TenantRepository;
 import com.flipmind.localizationservice.repositories.TranslatedStringRepository;
 import com.flipmind.localizationservice.repositories.TranslationRepository;
+import com.flipmind.localizationservice.utils.Utils;
 import com.wordnik.swagger.annotations.Api;
 import com.wordnik.swagger.annotations.ApiOperation;
 import com.wordnik.swagger.annotations.ApiParam;
@@ -62,6 +66,7 @@ public class TranslationController {
 	private TranslatedStringRepository translatedStringRepository;
 	
 	@RequestMapping(value = "/projects/{projectslug}/{documentslug}", method = RequestMethod.POST)
+	@ResponseBody
 	@ApiResponses(
 			value = {@ApiResponse(code = 404, message = "Something went wrong")
 	})
@@ -83,78 +88,86 @@ public class TranslationController {
 			HttpServletResponse response) throws ServletException, IOException {
 		try {
 		
-		String apiKey = request.getHeader(GlobalVariable.API_KEY);
-		
-		HttpHeaders headers = new HttpHeaders();
-		headers.add("Content-Type", "application/json; charset=utf-8");
-		
-		JSONMessage message = new JSONMessage();
-		message.setMessage("Adds a translation to the document");
-		
-		if (apiKey != null && !apiKey.isEmpty()) {
-			List<Tenant> tenants = tenantRepository.findByApiKey(apiKey);
+			String apiKey = request.getHeader(GlobalVariable.API_KEY);
 			
+			HttpHeaders headers = new HttpHeaders();
+			headers.add("Content-Type", "application/json; charset=utf-8");
 			
-			if (tenants != null && !tenants.isEmpty()) {
+			JSONMessage message = new JSONMessage();
+			message.setMessage("Adds a translation to the document");
+			
+			if (apiKey != null && !apiKey.isEmpty()) {
+				List<Tenant> tenants = tenantRepository.findByApiKey(apiKey);
 				
-				Translation translation = new Translation();
 				
-				//begin document data
-				if (jsonTranslation.getDocument() != null && jsonTranslation.getDocument().getId() > 0) {
-					Document document = documentRepository.findOne(jsonTranslation.getDocument().getId());
-					if (document != null) {
-						translation.setDocument(document);
-					} 
-				}
-				
-				if (translation.getDocument() == null) {
-					Tenant tenant = tenants.get(0);
-					for (Project project : tenant.getProjects()) {
-						
-						if (project.getSlug().equals(projectSlug)) {
-							for (Document document : project.getDocuments()) {
-								
-								if (document.getSlug().equals(documentSlug)) {
-									translation.setDocument(document);
+				if (tenants != null && !tenants.isEmpty()) {
+					
+					Translation translation = new Translation();
+					
+					//begin get document data
+					if (jsonTranslation.getDocument() != null && jsonTranslation.getDocument().getId() > 0) {
+						Document document = documentRepository.findOne(jsonTranslation.getDocument().getId());
+						if (document != null) {
+							translation.setDocument(document);
+						} 
+					}
+					
+					if (translation.getDocument() == null) {
+						Tenant tenant = tenants.get(0);
+						for (Project project : tenant.getProjects()) {
+							
+							if (project.getSlug().equals(projectSlug)) {
+								for (Document document : project.getDocuments()) {
+									
+									if (document.getSlug().equals(documentSlug)) {
+										translation.setDocument(document);
+									}
+									
 								}
-								
 							}
+							
+						}					
+					}
+					//end get document data
+					
+					//begin get locale Data
+					String localeCode = jsonTranslation.getLocale();
+					List<Locale> locales = localeRepository.findByLocaleCode(localeCode);
+					if (locales != null && !locales.isEmpty()) {
+						translation.setLocale(locales.get(0));
+					}
+					//end get locale Data
+					
+					//get status
+					Status status = Status.DRAFT;
+					if (jsonTranslation.getStatus() != null) {
+						if (jsonTranslation.getStatus().equalsIgnoreCase("PUBLISHED")) {
+							status = Status.PUBLISHED;
+						} else if (jsonTranslation.getStatus().equalsIgnoreCase("DELETED")) {
+							status = Status.DELETED;
 						}
-						
-					}					
+					}
+					translation.setStatus(status);
+					
+					//get active from date
+					translation.setActiveFromDate(Utils.getDateFromString(jsonTranslation.getActiveFrom()));
+					
+					boolean isSaveable = false;
+					if (translation.getDocument() != null && translation.getLocale() != null) {
+						isSaveable = true;
+					}
+					
+					if (isSaveable) {
+						translation = translationRepository.save(translation);
+						message.getItems().add("Translation: " + translation.getId() + " was added successfully");
+					}
 				}
-				//end get document data
+			} else {
 				
-				//begin get locale Data
-				String localeCode = jsonTranslation.getLocale();
-				List<Locale> locales = localeRepository.findByLocaleCode(localeCode);
-				if (locales != null && !locales.isEmpty()) {
-					translation.setLocale(locales.get(0));
-				}
-				//end get locale Data
-				
-				//get status
-				translation.setStatus(jsonTranslation.getStatus());
-				
-				//get active from date
-				translation.setActiveFromDate(new Date(jsonTranslation.getActiveFrom()));
-				
-				boolean isSaveable = false;
-				if (translation.getDocument() != null && translation.getLocale() != null) {
-					isSaveable = true;
-				}
-				
-				if (isSaveable) {
-					translation = translationRepository.save(translation);
-					message.getItems().add("Translation: " + translation.getId() + " was added successfully");
-				}
+				message.getItems().add("Request does not have authentication or path variables does not exist");
 			}
-		} else {
 			
-			message.getItems().add("Request does not have authentication or path variables does not exist");
-		}
-		
-		return new ResponseEntity<String>(new JSONSerializer().exclude("*.class").deepSerialize(message), headers, HttpStatus.OK); 
+			return new ResponseEntity<String>(new JSONSerializer().exclude("*.class").deepSerialize(message), headers, HttpStatus.OK); 
 		} catch (Exception e) {
 			e.printStackTrace();
 			request.getRequestDispatcher(GlobalVariable.ERROR_PATH).forward(request, response);
@@ -199,6 +212,9 @@ public class TranslationController {
 			
 			if (tenants != null && !tenants.isEmpty()) {
 				
+				//define a variable to check if there's no Translation in draft state
+				boolean isOneDraftTranslation = false;
+				
 				Tenant tenant = tenants.get(0);
 				for (Project project : tenant.getProjects()) {
 					
@@ -206,25 +222,73 @@ public class TranslationController {
 						for (Document document : project.getDocuments()) {
 							if (document.getSlug().equals(documentSlug)) {
 								List<Translation> translations = document.getTranslations();
+								
 								for (Translation translation : translations) {
-									
-									Locale locale = translation.getLocale();
-									if (locale != null && locale.getLocaleCode() != null && locale.getLocaleCode().equals(localeCode)) {
-										for (TranslatedString translatedString : translation.getTranslatedStrings()) {
-											if (stringKeys.contains(translatedString.getKey())) {
-												translatedStringRepository.delete(translatedString);
-												message.setSuccess(true);
-												message.getItems().add("TranslatedString " + translatedString.getKey() + " was deleted successfully");
+									System.out.println("====DEBUG===" + translation.getId());
+									//The Translation to be affected is the current one in the Draft state
+									if (translation.getStatus().equals(Status.DRAFT)) {
+										
+										isOneDraftTranslation = true;
+										Locale locale = translation.getLocale();
+										if (locale != null && locale.getLocaleCode() != null && locale.getLocaleCode().equals(localeCode)) {
+											for (TranslatedString translatedString : translation.getTranslatedStrings()) {
+												if (stringKeys.contains(translatedString.getKey())) {
+													translatedStringRepository.delete(translatedString);
+													message.setSuccess(true);
+													message.getItems().add("TranslatedString " + translatedString.getKey() + " was deleted successfully");
+												}
 											}
 										}
 									}
 								}
+								
+								//if none is in the draft status then a new Translation in the Draft status is created with 
+								//the strings copied form the previous Published version.
+								if (isOneDraftTranslation == false) {
+									System.out.println("====isOneDraftTranslation false ===");
+									for (Translation translation : translations) {
+										if (translation.getStatus().equals(Status.PUBLISHED)) {
+											System.out.println("====DEBUG: translation.getStatus().equals(Status.PUBLISHED)");
+											Translation newDraftTranslation = new Translation();
+											newDraftTranslation.setDocument(document);
+											newDraftTranslation.setStatus(Status.DRAFT);
+											List<Locale> locales = localeRepository.findByLocaleCode(localeCode);
+											if (locales != null && !locales.isEmpty()) {
+												newDraftTranslation.setLocale(locales.get(0));
+											}
+											newDraftTranslation = translationRepository.save(newDraftTranslation);
+											
+											//begin copy translated string to new translation
+											if (translation.getTranslatedStrings() != null) {
+												
+												List<TranslatedString> listToCopy = new ArrayList<TranslatedString>();
+												for (TranslatedString translatedString : translation.getTranslatedStrings()) {
+													TranslatedString copyTranslatedString = new TranslatedString();
+													copyTranslatedString.setKey(translatedString.getKey());
+													copyTranslatedString.setValue(translatedString.getValue());
+													copyTranslatedString.setTranslation(newDraftTranslation);
+													copyTranslatedString.setPlurals(translatedString.getPlurals());
+													copyTranslatedString = translatedStringRepository.save(copyTranslatedString);
+													listToCopy.add(copyTranslatedString);
+												}
+												
+												newDraftTranslation.setTranslatedStrings(listToCopy);
+											}
+											//end copy translated string to new translation
+											
+											message.getItems().add("There's no item in draft status. A new Translation in the draft status was created with id: " + newDraftTranslation.getId());
+										}
+										
+									}
+								}
+								break;
 							}
 						}
 					}
 				}
 				
-				if (message.isSuccess() == false) {
+				//print a message when translated strings can't be deleted via Draft Translation
+				if (message.isSuccess() == false && isOneDraftTranslation == true) {
 					message.getItems().add("Translated strings was not deleted, please recheck path variables");
 				}
 			} else {
